kullanıcı silinmesi sırasında silinmeyen tag sorunu devam ediyordu, çözüldü

+++ backend/src/main/java/com/application/backend/services/TagServiceImpl.java
@@ -1,6 +1,7 @@
 package com.application.backend.services;
 
 import java.util.List;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,36 +14,44 @@ import com.application.backend.repository.UserRepository;
 @Service
 public class TagServiceImpl implements TagService {
 
-  @Autowired TagRepository tagRepository;
-  @Autowired UserRepository userRepository;
+	@Autowired
+	TagRepository tagRepository;
+	@Autowired
+	UserRepository userRepository;
 
-  @Override
-  public void deleteTag(long tagId) {
-    tagRepository.deleteById(tagId);
-  }
+	@Override
+	public void deleteTag(long tagId) {
+		tagRepository.deleteById(tagId);
+	}
 
-  @Override
-  public List<Tags> ListImageTags(Images image) {
+	@Override
+	public List<Tags> ListImageTags(Images image) {
 
-    return tagRepository.findTagsByimages(image);
-  }
-  /*
-  @Override
-  public List<Tags> findTagsForUser() {
-    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-    String username = authentication.getName();
-    User user = userRepository.findByUsername(username);
+		return tagRepository.findTagsByimages(image);
+	}
 
-    return tagRepository.findAllByUser(user);
-  }*/
-  /*
-    @Override
-    public void deleteByUser(User user) {
-      Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-      String username = authentication.getName();
-      User user = userRepository.findByUsername(username);
-
-      tagRepository.deleteByUser(user);
-    }
-  */
+	@Override
+	public void deleteUserTag(long userId) {
+		/*
+		 * etiketi kullanan tek bir görsel var ise kullanıcı ile birlikte etiket de
+		 * siliniyor
+		 */
+		boolean deletable = false;
+		Set<Tags> userTags = tagRepository.findTagsByImages_User_Id(userId);
+		long tagId = 0;
+		for (Tags tag : userTags) {
+			tagId = tag.getId();
+			for (Images image : tag.getImages()) {
+				if (image.getUser().getId() == userId) {
+					deletable = true;
+				} else {
+					deletable = false;
+					break;
+				}
+			}
+			if (deletable) {
+				tagRepository.deleteById(tagId);
+			}
+		}
+	}
 }
